pieces connected outside of puzzle feature added
