+++ src/index.html
@@ -88,7 +88,7 @@
 
 
 <div class="instructions">
-  <p>Drag to place.</p>
+  <p>Drag to  or connect.</p>
   <p>Double click to rotate.</p>
 </div>
 <div id="container">
