+++ src/scripts/index.js
@@ -1,6 +1,7 @@
 import '../styles/index.scss';
 // import $ from 'jquery';
 import {pieceData} from './puzzlePieceHelper.js';
+import {checkSidePieces, getCoords} from './piecePlacementHelper.js';
 
 // console.log('webpack starterkit');
 
@@ -16,7 +17,7 @@ let idOfImageSelected;
 
 let director; //List with pieces and its data
 let offset;
-let pecaAlSeuLloc;
+let piecesPlacedCorrectly;
 let timeInitial;
 let timeEnd;
 
@@ -34,6 +35,7 @@ let puzzleImagesList = {};
 let imagesLoaded = 0;
 
 let piecesMatrix = [];
+let positionsList = [];
 
 const seeImage_btn = document.getElementById("seeImage");
 
@@ -101,6 +103,7 @@ function placePuzzleImagesInPlace() {
   window.puzzleImagesList = puzzleImagesList;
 
   piecesMatrix = [];
+  positionsList = [];
 
   let index = 0;
   for (let row=0; row<numRows; row++){
@@ -141,6 +144,8 @@ function placePuzzleImagesInPlace() {
       position.occupied=false;
       document.querySelector("#container").appendChild(position);
       position.appendChild(move);
+      positionsList.push(position);
+
       move.style.zIndex = zIndex++;
       move.zIndexPrevi=move.style.zIndex;
       move.x = col;
@@ -309,7 +314,7 @@ function getRandomImageId(max = 10) {
 function createPuzzle(){
   // imageSrc = `${host}/id/${getRandomImageId()}/${model.width}/${model.height}`;
 
-  pecaAlSeuLloc=0;
+  piecesPlacedCorrectly=0;
   // document.getElementById("container").innerHTML="";
   let child = document.getElementById("container").getElementsByClassName("position")[0];
   while (child) {
@@ -485,8 +490,13 @@ function start(){
     // e.style.top = getRandomInt(minHeightSpawnBottom, maxHeightSpawnBottom) + "px";
 
     // e.style.left=Math.random()*(window.innerWidth -100) +"px";
+
+
     setTimeout(function(){
       let angle=Math.floor(Math.random()*4);
+
+      // let angle=0;
+
       // e.setAttribute("classangle","g"+angle);
       e.rotation=angle*90;
       e.angle=angle;
@@ -552,27 +562,17 @@ function turn(){
 
 }
 
-function getCoords(elem) { // crossbrowser version
-  let box = elem.getBoundingClientRect();
-
-  let body = document.body;
-  let docEl = document.documentElement;
-
-  let scrollTop = window.pageYOffset || docEl.scrollTop || body.scrollTop;
-  let scrollLeft = window.pageXOffset || docEl.scrollLeft || body.scrollLeft;
-
-  let clientTop = docEl.clientTop || body.clientTop || 0;
-  let clientLeft = docEl.clientLeft || body.clientLeft || 0;
-
-  let top  = box.top +  scrollTop - clientTop;
-  let left = box.left + scrollLeft - clientLeft;
-
-  return { top: Math.round(top), left: Math.round(left) };
-}
-
 function fixPiece(p){
-  document.body.removeChild(p);
-  document.querySelectorAll(".position")[p.index].appendChild(p);
+  let correspondentPositionOnContainer = positionsList[p.index];
+  let offsetLeft = correspondentPositionOnContainer.offsetLeft + document.getElementById("container").offsetLeft;
+  let offsetTop = correspondentPositionOnContainer.offsetTop + document.getElementById("container").offsetTop;
+  p.style.top=offsetTop+"px";
+  p.style.left=offsetLeft+"px";
+  correspondentPositionOnContainer.occupied= true;
+  p.occupy = p.index;
+
+  p.parentElement.removeChild(p);
+  correspondentPositionOnContainer.appendChild(p);
   p.onmousedown="";
   p.onclick="";
   p.onmouseover="";
@@ -591,8 +591,8 @@ function fixPiece(p){
   setTimeout(function(){p.style.transform="scale(1)";},150);
   // setTimeout(function(){p.path.style.strokeWidth="2px"; p.style.zIndex= 0;},250);
   // setTimeout(function(){p.style.filter = "drop-shadow(black 0px 0px 0)"; p.style.zIndex= 0;},250);
-  pecaAlSeuLloc++;
-  if (pecaAlSeuLloc == numRows * numCols ){
+  piecesPlacedCorrectly++;
+  if (piecesPlacedCorrectly === numRows * numCols ){
     timeEnd = Math.floor(((new Date).getTime()-timeInitial)/1000);
     let hours = Math.floor(timeEnd/3600);
     let minutes =  Math.floor((timeEnd - hours * 3600)/60);
@@ -606,241 +606,98 @@ function fixPiece(p){
     setTimeout(final,500);
   }
 }
+
+
+/*
+*  Check if piece can be placed in correct position
+*/
 function placePiece(p){
   if(p.style) {
     p.style.zIndex = zIndex++;
-    p.zIndexPrevi=p.style.zIndex;
-    document.querySelectorAll(".position").forEach (function(e,i){
-      let offsetLeft = e.offsetLeft + document.getElementById("container").offsetLeft;
-      let offsetTop = e.offsetTop + document.getElementById("container").offsetTop;
-      if ((p.position().left>offsetLeft && p.position().left<offsetLeft+sizeOfPieces)&&(p.position().top>offsetTop && p.position().top<offsetTop+sizeOfPieces) && !e.occupied ){
-        p.style.left=offsetLeft+"px";
-        p.style.top=offsetTop+"px";
-        e.occupied= true;
-        p.occupy = i;
-        if (e.index == p.index && p.angle%4 === 0)fixPiece(p);
+    p.zIndexPrevi = p.style.zIndex;
+
+    let correspondentPositionOnContainer = positionsList[p.index];
+    let offsetLeft = correspondentPositionOnContainer.offsetLeft + document.getElementById("container").offsetLeft;
+    let offsetTop = correspondentPositionOnContainer.offsetTop + document.getElementById("container").offsetTop;
+    if ( ( p.position().left>offsetLeft && p.position().left<offsetLeft+sizeOfPieces ) &&
+        ( p.position().top>offsetTop && p.position().top<offsetTop+sizeOfPieces ) &&
+        !correspondentPositionOnContainer.occupied &&
+          correspondentPositionOnContainer.index === p.index && p.angle%4 === 0){
+        fixPiece(p);
+        if ( [...p.childNodes].find(child => child.className === 'move') ) {
+          let childrenToMove = [];
+          for (let i = 0; i <  p.childNodes.length ; i++) {
+            if( p.childNodes[i].className === 'move' ) {
+              childrenToMove.push(p.childNodes[i]);
+            }
+          }
+          childrenToMove.forEach(function(child){
+            fixPiece(child);
+          });
+        }
+        return true;
       }
-    });
+
+    // let positionsInContainer = document.querySelectorAll(".position");
+    // for (let i = 0; i < positionsInContainer.length; i++) {
+    //   let position = positionsInContainer[i];
+    //   let offsetLeft = position.offsetLeft + document.getElementById("container").offsetLeft;
+    //   let offsetTop = position.offsetTop + document.getElementById("container").offsetTop;
+    //   if ( ( p.position().left>offsetLeft && p.position().left<offsetLeft+sizeOfPieces ) &&
+    //     ( p.position().top>offsetTop && p.position().top<offsetTop+sizeOfPieces ) &&
+    //     !position.occupied &&
+    //     position.index === p.index && p.angle%4 === 0){
+    //     // if (position.index === p.index && p.angle%4 === 0) {
+    //     fixPiece(p, position, offsetTop, offsetLeft);
+    //     if ( [...p.childNodes].find(child => child.className === 'move') ) {
+    //       for (let i = 0; i <  p.childNodes.length ; i++) {
+    //         if( p.childNodes[i].className === 'move' ) {
+    //           fixPiece(p, position, offsetTop, offsetLeft);
+    //         }
+    //       }
+    //     }
+    //     return true;
+    //     // }
+    //   }
+    // }
+
+    // document.querySelectorAll(".position").forEach (function(e,i){
+    //
+    // });
+    return false;
   }
 }
 function dropPiece(){
+  let pieceToPlace = selectedPiece;
+  if( pieceToPlace && pieceToPlace.parentElement.className === 'move' ) pieceToPlace = pieceToPlace.parentElement;
+  let didItPlace = placePiece(pieceToPlace);
 
-  placePiece(selectedPiece);
+  console.log('did it placed', didItPlace);
 
-  checkSidePieces(selectedPiece);
-
+  if( !didItPlace ) {
+    // check if the piece can be connected
+    let hasConnected = checkSidePieces(selectedPiece, piecesMatrix, sizeOfPieces, numRows, numCols );
+    // if not, check its children
+    if(!hasConnected) {
+      let pieceToCheckChildren;
+      if( selectedPiece && selectedPiece.parentElement.className === 'move' ) {
+        pieceToCheckChildren = selectedPiece.parentElement;
+      } else if( selectedPiece && [...selectedPiece.childNodes].find(child => child.className === 'move') ) {
+        pieceToCheckChildren = selectedPiece;
+      }
+      if( pieceToCheckChildren ) {
+        for (let i = 0; i <  pieceToCheckChildren.childNodes.length ; i++) {
+          if( pieceToCheckChildren.childNodes[i].className === 'move' ) {
+            hasConnected = checkSidePieces(pieceToCheckChildren.childNodes[i], piecesMatrix, sizeOfPieces, numRows, numCols );
+            if(hasConnected) break;
+          }
+        }
+      }
+    }
+  }
 
   selectedPiece=false;
 
-}
-
-function checkSidePieces(selectedPiece) {
-
-  // console.log(selectedPiece[0]);
-
-  let threshold = 5;
-
-  // CHECK TOP PIECE
-  if( selectedPiece.y > 0 ) {
-    let pieceOnTop = piecesMatrix[selectedPiece.y-1][selectedPiece.x];
-    if( isSameAngle(selectedPiece, pieceOnTop) ) {
-      let isTopPieceNear = isOnTop(selectedPiece, pieceOnTop, threshold);
-      if( isTopPieceNear && isHierarchyAvailableToConnect(selectedPiece, pieceOnTop) ) {
-        checkHierarchyBeforeConnecting(selectedPiece, pieceOnTop, -sizeOfPieces, 0);
-        console.log('connect with piece in top: ', isTopPieceNear);
-        return;
-      }
-    }
-  }
-  // CHECK BOTTOM PIECE
-  if(  selectedPiece.y < numRows-1 ) {
-    let pieceOnBottom = piecesMatrix[selectedPiece.y+1][selectedPiece.x];
-    if( isSameAngle(selectedPiece, pieceOnBottom) ) {
-      let isBottomPieceNear = isOnTop(pieceOnBottom, selectedPiece, threshold);
-      if( isBottomPieceNear && isHierarchyAvailableToConnect(selectedPiece, pieceOnBottom) ) {
-        checkHierarchyBeforeConnecting(selectedPiece, pieceOnBottom, sizeOfPieces, 0);
-        console.log('connect with piece in bottom: ', isBottomPieceNear);
-        return;
-      }
-    }
-  }
-  // CHECK LEFT PIECE
-  if(  selectedPiece.x > 0 ) {
-    let pieceOnLeft = piecesMatrix[selectedPiece.y][selectedPiece.x - 1];
-    if( isSameAngle(selectedPiece, pieceOnLeft) ) {
-      let isLeftPieceNear = isOnRight(pieceOnLeft, selectedPiece, threshold);
-      if( isLeftPieceNear && isHierarchyAvailableToConnect(selectedPiece, pieceOnLeft) ) {
-        checkHierarchyBeforeConnecting(selectedPiece, pieceOnLeft, 0, -sizeOfPieces);
-        console.log('connect with piece in left: ', isLeftPieceNear);
-        return;
-      }
-    }
-  }
-  // CHECK RIGHT PIECE
-  if(  selectedPiece.x < numCols - 1 ) {
-    let pieceOnRight = piecesMatrix[selectedPiece.y][selectedPiece.x + 1];
-    if( isSameAngle(selectedPiece, pieceOnRight) ) {
-      let isRightPieceNear = isOnRight(selectedPiece, pieceOnRight, threshold);
-      if( isRightPieceNear && isHierarchyAvailableToConnect(selectedPiece, pieceOnRight) ) {
-        checkHierarchyBeforeConnecting(selectedPiece, pieceOnRight, 0, sizeOfPieces);
-        console.log('connect with piece in right: ', isRightPieceNear);
-        return;
-      }
-    }
-  }
-
-}
-
-function isSameAngle(piece1, piece2) {
-  return piece1.angle === piece2.angle;
-}
-
-function isOnTop(bottomPiece, topPiece, threshold) {
-  // let topRect = bottomPiece.offsetTop - threshold;
-  // let bottomRect = bottomPiece.offsetTop + threshold;
-  // let rightRect = bottomPiece.offsetLeft + parseInt(bottomPiece.clientWidth / 2) + threshold;
-  // let leftRect = bottomPiece.offsetLeft + parseInt(bottomPiece.clientWidth / 2) - threshold;
-  //
-  // let pieceOnTopTopRect = topPiece.offsetTop + topPiece.clientHeight - threshold;
-  // let pieceOnTopBottomRect = topPiece.offsetTop + topPiece.clientHeight + threshold;
-  // let pieceOnTopRightRect = topPiece.offsetLeft + parseInt(topPiece.clientWidth / 2) + threshold;
-  // let pieceOnTopLeftRect = topPiece.offsetLeft + parseInt(topPiece.clientWidth / 2) - threshold;
-
-  let bottomPieceCoords = getCoords(bottomPiece);
-  let topRect = bottomPieceCoords.top - threshold;
-  let bottomRect = bottomPieceCoords.top + threshold;
-  let rightRect = bottomPieceCoords.left + parseInt(bottomPiece.clientWidth / 2) + threshold;
-  let leftRect = bottomPieceCoords.left + parseInt(bottomPiece.clientWidth / 2) - threshold;
-  let pieceOnTopCoords = getCoords(topPiece);
-  let pieceOnTopTopRect = pieceOnTopCoords.top + topPiece.clientHeight - threshold;
-  let pieceOnTopBottomRect = pieceOnTopCoords.top + topPiece.clientHeight + threshold;
-  let pieceOnTopRightRect = pieceOnTopCoords.left + parseInt(topPiece.clientWidth / 2) + threshold;
-  let pieceOnTopLeftRect = pieceOnTopCoords.left + parseInt(topPiece.clientWidth / 2) - threshold;
-
-  return rectanglesIntersect(leftRect,topRect,rightRect,bottomRect , pieceOnTopLeftRect,pieceOnTopTopRect,pieceOnTopRightRect,pieceOnTopBottomRect );
-}
-
-function isOnRight(leftPiece, rightPiece, threshold) {
-  // let topRect = leftPiece.offsetTop + parseInt(leftPiece.clientHeight / 2) - threshold;
-  // let bottomRect = leftPiece.offsetTop + parseInt(leftPiece.clientHeight / 2) + threshold;
-  // let rightRect = leftPiece.offsetLeft + leftPiece.clientWidth + threshold;
-  // let leftRect = leftPiece.offsetLeft + leftPiece.clientWidth - threshold;
-  //
-  // let pieceOnRightTopRect = rightPiece.offsetTop + parseInt(rightPiece.clientHeight / 2) - threshold;
-  // let pieceOnRightBottomRect = rightPiece.offsetTop + parseInt(rightPiece.clientHeight / 2) + threshold;
-  // let pieceOnRightRightRect = rightPiece.offsetLeft + threshold;
-  // let pieceOnRightLeftRect = rightPiece.offsetLeft - threshold;
-
-  let leftPieceCoords = getCoords(leftPiece);
-  let topRect = leftPieceCoords.top + parseInt(leftPiece.clientHeight / 2) - threshold;
-  let bottomRect = leftPieceCoords.top + parseInt(leftPiece.clientHeight / 2) + threshold;
-  let rightRect = leftPieceCoords.left + leftPiece.clientWidth + threshold;
-  let leftRect = leftPieceCoords.left + leftPiece.clientWidth - threshold;
-  let rightPieceCoords = getCoords(rightPiece);
-  let pieceOnRightTopRect = rightPieceCoords.top + parseInt(rightPiece.clientHeight / 2) - threshold;
-  let pieceOnRightBottomRect = rightPieceCoords.top + parseInt(rightPiece.clientHeight / 2) + threshold;
-  let pieceOnRightRightRect = rightPieceCoords.left + threshold;
-  let pieceOnRightLeftRect = rightPieceCoords.left - threshold;
-
-  return rectanglesIntersect(leftRect,topRect,rightRect,bottomRect , pieceOnRightLeftRect,pieceOnRightTopRect,pieceOnRightRightRect,pieceOnRightBottomRect );
-}
-
-function rectanglesIntersect(
-  minAx, minAy, maxAx, maxAy,
-  minBx, minBy, maxBx, maxBy ) {
-  return maxAx >= minBx && minAx <= maxBx && minAy <= maxBy && maxAy >= minBy
-}
-
-function isHierarchyAvailableToConnect(pieceToConnectTo, pieceToBeConnected) {
-  if(pieceToConnectTo.contains(pieceToBeConnected) || pieceToBeConnected.contains(pieceToConnectTo)) return false;
-  if(pieceToConnectTo.parentElement.className === 'move' && pieceToConnectTo.parentElement === pieceToBeConnected.parentElement) return false;
-  return true;
-}
-
-function checkHierarchyBeforeConnecting(pieceToConnectTo, pieceToBeConnected, topPosition, leftPosition) {
-  // topPosition = (pieceToBeConnected.y - selectedPiece.y) * sizeOfPieces;
-  // leftPosition = (pieceToBeConnected.x - selectedPiece.x) * sizeOfPieces;
-
-  //check if piece to be connected is a parent, in that case the parent piece and all the children have to pass for the piece to connect to
-  if([...pieceToBeConnected.childNodes].find(child => child.className === 'move')) {
-    console.log('is a parent, children will be moved to piece to connect. Children: ', pieceToBeConnected.childNodes);
-    connectChildPieces(pieceToConnectTo, pieceToBeConnected);
-    connectPieces(pieceToConnectTo, pieceToBeConnected);
-  }
-  // check if the piece to be connected has a parent piece, in that case connect the parent and all its children
-  else if(pieceToBeConnected.parentElement.className === 'move' ) {
-    console.log('is child, connect parent');
-    pieceToBeConnected = pieceToBeConnected.parentElement;
-
-    //if destination piece has a parent piece, connect to parent piece instead
-    if( pieceToConnectTo.parentElement.className === 'move' ) {
-      pieceToConnectTo = pieceToConnectTo.parentElement;
-    }
-
-    connectChildPieces(pieceToConnectTo, pieceToBeConnected);
-    connectPieces(pieceToConnectTo, pieceToBeConnected);
-  }else {
-    connectPieces(pieceToConnectTo, pieceToBeConnected);
-  }
-
-
-}
-
-function connectChildPieces(pieceToConnectTo, pieceToBeConnected) {
-  console.log('children', pieceToBeConnected.children);
-  let childrenToMove = [];
-
-  pieceToBeConnected.childNodes.forEach(function(child){
-    if(child.className === 'move') {
-      childrenToMove.push(child);
-      // console.log('connect child', child);
-    }
-  });
-
-  childrenToMove.forEach(function(child){
-    connectPieces(pieceToConnectTo, child);
-  });
-
-}
-
-function connectPieces (pieceToConnectTo, pieceToBeConnected){
-    // if(pieceToConnectTo.contains(pieceToBeConnected) || pieceToBeConnected.contains(pieceToConnectTo)) return;
-  let topPosition = (pieceToBeConnected.y - pieceToConnectTo.y) * sizeOfPieces;
-  let leftPosition = (pieceToBeConnected.x - pieceToConnectTo.x) * sizeOfPieces;
-  // //check if piece to be connected is a parent, in that case the parent piece and all the children have to pass for the piece to connect to
-  // if([...pieceToBeConnected.childNodes].find(child => child.className === 'move')) {
-  //   console.log('is a parent, children will be moved to piece to connect');
-  //   pieceToBeConnected.childNodes.forEach(function(child){
-  //     if(child.className === 'move') {
-  //       console.log('pieceToConnectTo', pieceToConnectTo);
-  //       console.log('child', child);
-  //       console.trace();
-  //       connectPieces(pieceToConnectTo, child);
-  //     }
-  //   });
-  //   // connectPieces(pieceToConnectTo, child);
-  // }
-  //
-  // if(pieceToBeConnected.parentElement.className === 'move' ) {
-  //   connectPieces(pieceToConnectTo, pieceToBeConnected.parentElement);
-  //   return;
-  // }
-
-  // if(pieceToConnectTo.parentElement.className === 'move') {
-  //   topPosition = (pieceToBeConnected.y - selectedPiece.parentElement.y) * sizeOfPieces;
-  //   leftPosition = (pieceToBeConnected.x - selectedPiece.parentElement.x) * sizeOfPieces;
-  //   pieceToConnectTo = pieceToConnectTo.parentElement;
-  // }
-
-
-  console.log('pieceToConnectTo', pieceToConnectTo);
-  console.log('pieceToBeConnected', pieceToBeConnected);
-  console.log('parent element', pieceToBeConnected.parentElement);
-
-  pieceToBeConnected.parentElement.removeChild(pieceToBeConnected);
-  pieceToConnectTo.appendChild(pieceToBeConnected);
-  pieceToBeConnected.style.top = topPosition + 'px';
-  pieceToBeConnected.style.left = leftPosition + 'px';
 }
 
 document.addEventListener('mousemove', drag, {passive: false});
@@ -882,7 +739,7 @@ function drag(e) {
 
 function takePiece(e){
 
-  console.log('take piece')
+  // console.log('take piece')
 
   if(e.target.parentElement.className === 'position') {
     return;
