+++ src/styles/index.scss
@@ -408,7 +408,7 @@ h2{
 }
 
 .radio-toolbar input[type="radio"]:checked + label {
-  background-color: #52a067;
+  background-color: #444444;
   border-color: #dddddd;
   color: #fff;
 }
@@ -435,7 +435,7 @@ h2{
 .instructions p {
   margin: 0;
   display: inline-block;
-  color: #8a8a8a;
+  color: #ffffff;
 }
 
 #arrangePieces, #seeImage {
